Refactor docs module
This is a total refactor of the previous docs module. This switches the
TestDocSet class (previously BenchmarkTestDocSet) so that making docs is
done via generators. It also implements a new FileSet class for
persisting generated data to disk, where JSON objects are streamed to a
file, to support much larger data sets. Timing tests show this works
pretty efficiently, at least so far.