Make QOL improvements and update README
I've made another pass at the `docs` module, looking at how docsets are
generated and used and how they interact with FileSets. This adds some
QOL improvements, particularly around reusing DocSets generated from
schemas for multiple tests. If the schema is saved to disk, a DocSet
will now seamlessly switch to reading from the saved file for the next
iteration -- the user doesn't have to manage this. This required a bit
of reengineering.

I also went through the extended example in the README and tried to make
sure everything matches all of the refactoring.

Finally, I updated pyproject.toml to make pysolr a normal dependency and
not just a test dependency. Though this package only calls pysolr during
tests, it's still a requirement for people using the package.