Update schema for solrfixtures.profile update
I made some minor updates to solrfixtures.profile to implement hidden
fields, which changed both the profile.Field.__init__ signature and the
profile.Schema.add_fields method. This updates the subclasses of those
objects in the `schema` module to match. I've also modified
BenchmarkSchema so that it actually calls the parent (i.e., in __init__
and add_fields) instead of duplicating the parent functionality.