Fix error when injecting search terms: SearchField
During injection, an RNG call to `choice` is used to pick the position
in a string where terms should be injected. There was an off-by-one
error making it possible to submit an empty set of choices to `choice`
when injecting into a single-character string. This fixes that error and
also adds tests for a few related edge-cases.