Add arg to search methods to specify blank search
Depending on how the target Solr collection is configured and what query
parser is being used, the needed `q` value to search a blank value and
return all docs may vary. I've added a `blank_q` value to convert an
empty q to a search value that will return all documents. (Such as
'*:*'.)