Update integration test to specify `blank_q`
The previous changes added the ability to specify the required `q` value
for performing "blank" searches, e.g. when using the `run_searches`
method of BenchmarkRunner. Apparently I neglected to update the runner
integration test (`tests/test_runner_integration.py`) with the new
parameter to allow it continue to pass. This fixes the problem.