Fix invalid kwarg error on Python 3.7 (unlink)
The FileSet `clear` method uses the Path `unlink` method to delete
files. On Python 3.8 and above, `unlink` has a kwarg `missing_ok` which
lets you silence the error raised when trying to delete a non-existent
file. On Python 3.7, `unlink` does not have this kwarg. To fix this, we
just have to use a try/except block to catch (and ignore) the
FileNotFoundError.

(I didn't catch this until now because I didn't have the tox
configuration in place to check multiple Python versions.)