Finalize type hints and add mypy check to tox cfg
I added mypy --strict checks to the tox config to run along with the
other linters. As expected, running mypy revealed a lot of corrections
to make, which I've done. Because type hints are now more or less
correct, I've also added a `py.typed` file to the package.

There are a handful of small changes I've made to the code to make mypy
happy, nothing major.

Note: The additional mypy flags in the tox environment are necessary
because `fauxdoc` is not yet flagged as being typed. When I get a chance
to run mypy on that package, I will add `py.typed` there and release a
new version. At that point I can remove those flags here and rerun mypy
to fix any lingering issues.